worked on review comment suggestions
